+++ packages/expo-sqlite/src/SQLite.ts
@@ -2,12 +2,19 @@ import './polyfillNextTick';
 
 import customOpenDatabase from '@expo/websql/custom';
 import { NativeModulesProxy } from 'expo-modules-core';
-import zipObject from 'lodash/zipObject';
 import { Platform } from 'react-native';
 
-import { Query, SQLiteCallback, ResultSet, ResultSetError, WebSQLDatabase } from './SQLite.types';
+import { Query, ResultSet, ResultSetError, SQLiteCallback, WebSQLDatabase } from './SQLite.types';
 
 const { ExponentSQLite } = NativeModulesProxy;
+
+function zipObject(keys: string[], values: any[]) {
+  const result = {};
+  for (let i = 0; i < keys.length; i++) {
+    result[keys[i]] = values[i];
+  }
+  return result;
+}
 
 class SQLiteDatabase {
   _name: string;